+++ relay-config/src/config.rs
@@ -762,6 +762,10 @@ pub enum KafkaTopic {
     Sessions,
     /// Aggregate Metrics.
     Metrics,
+    /// Profiling sessions
+    ProfilingSessions,
+    /// Profiling traces
+    ProfilingTraces,
 }
 
 /// Configuration for topics.
@@ -782,6 +786,10 @@ pub struct TopicAssignments {
     pub sessions: TopicAssignment,
     /// Metrics topic name.
     pub metrics: TopicAssignment,
+    /// Profiling sessions topic name
+    pub profiling_sessions: TopicAssignment,
+    /// Profiling traces topic name
+    pub profiling_traces: TopicAssignment,
 }
 
 impl TopicAssignments {
@@ -795,6 +803,8 @@ impl TopicAssignments {
             KafkaTopic::OutcomesBilling => self.outcomes_billing.as_ref().unwrap_or(&self.outcomes),
             KafkaTopic::Sessions => &self.sessions,
             KafkaTopic::Metrics => &self.metrics,
+            KafkaTopic::ProfilingSessions => &self.profiling_sessions,
+            KafkaTopic::ProfilingTraces => &self.profiling_traces,
         }
     }
 }
@@ -809,6 +819,8 @@ impl Default for TopicAssignments {
             outcomes_billing: None,
             sessions: "ingest-sessions".to_owned().into(),
             metrics: "ingest-metrics".to_owned().into(),
+            profiling_sessions: "profiling-sessions".to_owned().into(),
+            profiling_traces: "profiling-traces".to_owned().into(),
         }
     }
 }